backend/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
mixer==6.1.3
django-cors-headers==3.2.1